+++ db/migrate/20170315194830_rename_required_to_requirement_for_design_options.rb
@@ -1,0 +1,5 @@
+class RenameRequiredToRequirementForDesignOptions < ActiveRecord::Migration[5.0]
+  def change
+    rename_column :design_options, :required, :requirement
+  end
+end